us_backend/requirements.txt
fastapi
uvicorn
python-dotenv
alpaca-py
requests
beautifulsoup4
aiosqlite
pandas
pandas-ta
aiohttp
